requirements.txt
fastapi
uvicorn[standard]
pydantic
mysql-connector-python
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
